More work on the emulator
It seems that we cannot do (~a | bit) to check if a bit is not set.
Maybe this only happens with unsigned types?

The emulator seems to be running some kind of loop of instruction. There
still couple of thing to implement such as the stack and prefix cb stuff.