Extract information from to show to the user
We also changed the getopt to allow long. --info will dump ROM
information to the user after loading the ROM and --emulate and --disassemble options are also
enabled. Right now we can only load ROM's that use only single ROM Bank,
so we disabled to the options to run ROM's with multiple ROM Banks.